requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile -U requirements.in
#
-e git+https://github.com/tiltec/django-influxdb-metrics@07f5701#egg=django-influxdb-metrics
-e git+https://github.com/tiltec/django-rest-swagger@7a9ef64#egg=django-rest-swagger
aioredis==1.2.0           # via channels-redis
apipkg==1.5               # via execnet
argon2-cffi==19.1.0       # via django
asgiref==2.3.2            # via channels, channels-redis
async-timeout==3.0.1      # via aioredis, asgiref
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via automat, pytest, twisted
autobahn==19.3.3          # via daphne
autoflake==1.3
automat==0.7.0            # via twisted
autopep8==1.4.4
babel==2.6.0
backcall==0.1.0           # via ipython
bleach-whitelist==0.0.10
bleach==3.1.0
cchardet==2.1.4           # via talon
certifi==2019.3.9         # via requests
cffi==1.12.3              # via argon2-cffi
channels==2.1.7
channels_redis==2.3.2
chardet==3.0.4            # via requests, talon
click==7.0                # via pip-tools
constantly==15.1.0        # via twisted
coreapi==2.3.3            # via openapi-codec
coreschema==0.0.4         # via coreapi
coverage==4.5.3           # via pytest-cov
cssselect==1.0.3          # via talon
daphne==2.2.5
decorator==4.4.0          # via ipython, traitlets
django-anymail==6.0
django-cors-headers==2.5.3
django-crispy-forms==1.7.2
django-dirtyfields==1.3.1
django-enumfield==1.6b3
django-extensions==2.1.6
django-filter==2.1.0
django-jinja==2.4.1
django-redis==4.10.0
django-silk==3.0.2
django-timezone-field==3.0
django-versatileimagefield==1.10
django[argon2]==2.2.1
djangorestframework==3.9.3
entrypoints==0.3          # via flake8
execnet==1.6.0            # via pytest-xdist
factory-boy==2.11.1
faker==1.0.5              # via factory-boy
flake8==3.7.7
freezegun==0.3.11
furl==2.0.0
gevent==1.4.0
gprof2dot==2017.9.19      # via django-silk
greenlet==0.4.15          # via gevent
hiredis==1.0.0
html2text==2018.1.9
html5lib==1.0.1           # via talon
huey==1.11.0
hyperlink==19.0.0         # via twisted
idna==2.8                 # via hyperlink, requests
incremental==17.5.0       # via twisted
influxdb==5.2.2
ipython-genutils==0.2.0   # via traitlets
ipython==7.5.0
itypes==1.1.0             # via coreapi
jedi==0.13.3              # via ipython
jinja2==2.10.1            # via coreschema, django-jinja, django-silk
logging-tree==1.8
lxml==4.3.3               # via talon
markdown==3.1
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
more-itertools==7.0.0
msgpack==0.5.6            # via channels-redis
numpy==1.16.3             # via scikit-learn, scipy, talon
openapi-codec==1.3.2
orderedmultidict==1.0     # via furl
parso==0.4.0              # via jedi
pathlib2==2.3.3
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==5.4.1             # via django-versatileimagefield
pip-tools==3.6.1
pluggy==0.9.0             # via pytest
prompt-toolkit==2.0.9     # via ipython
psutil==5.6.1             # via python-server-metrics
psycopg2-binary==2.8.2
ptyprocess==0.6.0         # via pexpect
py==1.8.0                 # via pytest
pycodestyle==2.5.0        # via autopep8, flake8
pycparser==2.19           # via cffi
pyfcm==1.4.7
pyflakes==2.1.1           # via autoflake, flake8
pygments==2.3.1           # via django-silk, ipython
pyhamcrest==1.9.0         # via twisted
pymdown-extensions==6.0
pytest-asyncio==0.10.0
pytest-cov==2.6.1
pytest-django==3.4.8
pytest-forked==1.0.2      # via pytest-xdist
pytest-xdist==1.28.0
pytest==4.4.1             # via pytest-asyncio, pytest-cov, pytest-django, pytest-forked, pytest-xdist
python-dateutil==2.8.0    # via django-silk, faker, freezegun, influxdb
python-server-metrics==0.2.1
pytz==2019.1
raven==6.10.0
redis==3.2.1
regex==2019.4.14          # via talon
requests-mock==1.6.0
requests==2.21.0          # via coreapi, django-anymail, django-silk, influxdb, pyfcm, requests-mock, sparkpost
scikit-learn==0.20.3      # via talon
scipy==1.2.1              # via scikit-learn, talon
simplejson==3.16.0
six==1.12.0               # via argon2-cffi, autobahn, automat, bleach, django-anymail, django-extensions, faker, freezegun, furl, html5lib, influxdb, orderedmultidict, pathlib2, pip-tools, prompt-toolkit, pyhamcrest, pytest, pytest-xdist, python-dateutil, requests-mock, talon, tld, traitlets, txaio
sparkpost==1.3.6
sqlparse==0.3.0           # via django, django-silk
talon==1.4.4
tblib==1.4.0
text-unidecode==1.2       # via faker
tld==0.9.3
traitlets==4.3.2          # via ipython
twisted==19.2.0           # via daphne
txaio==18.8.1             # via autobahn
uritemplate==3.0.0        # via coreapi
urllib3==1.24.2           # via requests
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via bleach, html5lib
yapf==0.27.0
zope.interface==4.6.0     # via twisted
